# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trip API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get all trips",
                "operationId": "692b0f6b729af25639ef76046f8fba1f",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Filter by starting point",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Filter by ending point",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by starting date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of trips",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trip"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Create a new trip",
                "operationId": "330c41008c88334c0e7653cbe4f0710a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trip created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get a trip by ID",
                "operationId": "6b8785740e1e7723d8899ea7bd468b48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the trip to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Trip"
                ],
                "summary": "Update a trip",
                "operationId": "92f9255028e296ba3829d9f96d7c37ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the trip to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TripRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trip updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trip"
                ],
                "summary": "Delete a trip",
                "operationId": "ec4541f129cf6d7a16feb2989476dd3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the trip to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Trip deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Trip": {
                "title": "Trip",
                "required": [
                    "starting_point",
                    "ending_point",
                    "starting_at",
                    "available_places",
                    "price"
                ],
                "properties": {
                    "id": {
                        "description": "Trip ID",
                        "type": "integer"
                    },
                    "starting_point": {
                        "description": "Starting point",
                        "type": "string"
                    },
                    "ending_point": {
                        "description": "Ending point",
                        "type": "string"
                    },
                    "starting_at": {
                        "description": "Starting date",
                        "type": "string",
                        "format": "date"
                    },
                    "available_places": {
                        "description": "Number of available places",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Price of the trip",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID of the user who created the trip",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "TripRequest": {
                "title": "TripRequest",
                "required": [
                    "starting_point",
                    "ending_point",
                    "starting_at",
                    "available_places",
                    "price"
                ],
                "properties": {
                    "starting_point": {
                        "description": "Starting point",
                        "type": "string"
                    },
                    "ending_point": {
                        "description": "Ending point",
                        "type": "string"
                    },
                    "starting_at": {
                        "description": "Starting date",
                        "type": "string",
                        "format": "date"
                    },
                    "available_places": {
                        "description": "Number of available places",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Price of the trip",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}